refactor(P_IMPLEMENTATION_SP_C8_REFACTOR): Implement interface extraction and package facade re-exports (#406)
