+++ .phase-gate/state.json
@@ -88,5 +88,5 @@
     }
   ],
   "reconstructed": false,
-  "current_sub_phase": null
+  "current_sub_phase": "red"
 }
